move payload to own module
