+++ meilisearch-http/src/extractors/mod.rs
@@ -1,0 +1,1 @@
+pub mod payload;
