+++ meilisearch-http/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod data;
 #[macro_use]
 pub mod error;
+pub mod extractors;
 pub mod helpers;
 mod index;
 mod index_controller;
@@ -10,19 +11,12 @@ pub mod routes;
 #[cfg(all(not(debug_assertions), feature = "analytics"))]
 pub mod analytics;
 
-use std::{
-    pin::Pin,
-    task::{Context, Poll},
-};
-
 pub use self::data::Data;
-use futures::{
-    future::{ready, Ready},
-    Stream,
-};
 pub use option::Opt;
 
-use actix_web::{dev, error::PayloadError, web, FromRequest, HttpRequest};
+use actix_web::web;
+
+use extractors::payload::PayloadConfig;
 
 pub fn configure_data(config: &mut web::ServiceConfig, data: Data) {
     let http_payload_size_limit = data.http_payload_size_limit();
@@ -113,66 +107,4 @@ macro_rules! create_app {
                 middleware::TrailingSlash::Trim,
             ))
     }};
-}
-
-pub struct Payload {
-    payload: dev::Payload,
-    limit: usize,
-}
-
-pub struct PayloadConfig {
-    limit: usize,
-}
-
-impl PayloadConfig {
-    pub fn new(limit: usize) -> Self {
-        Self { limit }
-    }
-}
-
-impl Default for PayloadConfig {
-    fn default() -> Self {
-        Self { limit: 256 * 1024 }
-    }
-}
-
-impl FromRequest for Payload {
-    type Config = PayloadConfig;
-
-    type Error = PayloadError;
-
-    type Future = Ready<Result<Payload, Self::Error>>;
-
-    #[inline]
-    fn from_request(req: &HttpRequest, payload: &mut dev::Payload) -> Self::Future {
-        let limit = req
-            .app_data::<PayloadConfig>()
-            .map(|c| c.limit)
-            .unwrap_or(Self::Config::default().limit);
-        ready(Ok(Payload {
-            payload: payload.take(),
-            limit,
-        }))
-    }
-}
-
-impl Stream for Payload {
-    type Item = Result<web::Bytes, PayloadError>;
-
-    #[inline]
-    fn poll_next(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
-        match Pin::new(&mut self.payload).poll_next(cx) {
-            Poll::Ready(Some(result)) => match result {
-                Ok(bytes) => match self.limit.checked_sub(bytes.len()) {
-                    Some(new_limit) => {
-                        self.limit = new_limit;
-                        Poll::Ready(Some(Ok(bytes)))
-                    }
-                    None => Poll::Ready(Some(Err(PayloadError::Overflow))),
-                },
-                x => Poll::Ready(Some(x)),
-            },
-            otherwise => otherwise,
-        }
-    }
 }
